requirements.txt
scapy>=2.5.0
paramiko>=3.4.0
pyyaml>=6.0
jinja2>=3.1.2
rich>=13.7.0
click>=8.1.7
numpy>=1.24.0
